server/requirements.txt
requests
fastapi
uvicorn
argon2-cffi
